Minor cleanups (#9743)
* Minor cleanups

* Delete slasher client files

* Revert "Delete slasher client files"

This reverts commit 0c995a1d4a834ded97385e6e338fe630c81db26b.

* Update slasher_client.go

+++ beacon-chain/p2p/peers/scorers/service_test.go
@@ -143,7 +143,7 @@ func TestScorers_Service_Score(t *testing.T) {
 		s, pids := setupScorer()
 		// Peers start with boosted start score (new peers are boosted by block provider).
 		startScore := float64(0)
-		penalty := (-10 / float64(s.BadResponsesScorer().Params().Threshold)) * float64(0.3)
+		penalty := (-10 / float64(s.BadResponsesScorer().Params().Threshold)) * 0.3
 
 		// Update peers' stats and test the effect on peer order.
 		s.BadResponsesScorer().Increment("peer2")
@@ -197,7 +197,7 @@ func TestScorers_Service_Score(t *testing.T) {
 		s, _ := setupScorer()
 		s1 := s.BlockProviderScorer()
 		s2 := s.BadResponsesScorer()
-		penalty := (-10 / float64(s.BadResponsesScorer().Params().Threshold)) * float64(0.3)
+		penalty := (-10 / float64(s.BadResponsesScorer().Params().Threshold)) * 0.3
 
 		// Full score, no penalty.
 		s1.IncrementProcessedBlocks("peer1", batchSize*5)
